chore: Change names and create directories for different graphs
