+++ Assets/Scenes/Grid2D.unity
@@ -5628,7 +5628,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: af7ee45d9f0642d9bceca6fda72457ae, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _requestType: 0
+  _requestType: 1
 --- !u!1 &196024185
 GameObject:
   m_ObjectHideFlags: 0
