+++ Assets/Scripts/NavigationGraph/Graph/Planar/Jobs/CountNeighborsJob.cs
@@ -3,16 +3,15 @@ using Unity.Collections;
 using Unity.Jobs;
 using Unity.Mathematics;
 
-namespace NavigationGraph.Graph
+namespace NavigationGraph.Graph.Planar.Jobs
 {
+    // It could be separated in three diferent jobs, so I dont't have to ask all the time the NeigbhorsPerCell size.
     [BurstCompile]
-    internal struct PrecomputeNeighborsJob : IJob
+    internal struct CountNeighborsJob : IJob
     {
         [ReadOnly] public NativeArray<Node> grid;
         [ReadOnly] public NativeArray<int2> offsets16;
-        [ReadOnly] public NativeArray<int> neighborCounts;
-        [ReadOnly] public NativeArray<int> neighborOffsets;
-        [WriteOnly] public NativeArray<int> allNeighbors;
+        [WriteOnly] public NativeArray<int> neighborCounts;
         public int gridSizeX;
         public int gridSizeZ;
         public float maxHeightDifference;
@@ -23,9 +22,6 @@ namespace NavigationGraph.Graph
             for (int i = 0; i < grid.Length; i++)
             {
                 int count = 0;
-                int baseIndex = neighborOffsets[i];
-                Node currentCell = grid[i];
-                int maxNeighbors = neighborCounts[i];
 
                 if (neighborsPerCell != NeighborsPerCell.Sixteen)
                 {
@@ -34,7 +30,6 @@ namespace NavigationGraph.Graph
                     {
                         for (int offsetZ = -range; offsetZ <= range; offsetZ++)
                         {
-                            if (count >= maxNeighbors) break;
                             if (offsetX == 0 && offsetZ == 0) continue;
 
                             if (neighborsPerCell == NeighborsPerCell.Four &&
@@ -43,19 +38,15 @@ namespace NavigationGraph.Graph
                             if (neighborsPerCell == NeighborsPerCell.Eight &&
                                 math.max(math.abs(offsetX), math.abs(offsetZ)) > 1) continue;
 
-                            int gridX = currentCell.gridX + offsetX;
-                            int gridZ = currentCell.gridZ + offsetZ;
+                            int gridX = grid[i].gridX + offsetX;
+                            int gridZ = grid[i].gridZ + offsetZ;
 
                             if (gridX >= 0 && gridX < gridSizeX &&
                                 gridZ >= 0 && gridZ < gridSizeZ)
                             {
-                                int neighborIndex = gridX + gridZ * gridSizeX;
-                                var yDistance = math.abs(currentCell.position.y - grid[neighborIndex].position.y);
+                                var yDistance = grid[i].position.y - grid[gridX + gridZ * gridSizeX].position.y;
                                 if (yDistance < maxHeightDifference)
-                                {
-                                    allNeighbors[baseIndex + count] = neighborIndex;
                                     count++;
-                                }
                             }
                         }
                     }
@@ -64,24 +55,20 @@ namespace NavigationGraph.Graph
                 {
                     foreach (var offset in offsets16)
                     {
-                        if (count >= maxNeighbors) break;
-
-                        int gridX = currentCell.gridX + offset.x;
-                        int gridZ = currentCell.gridZ + offset.y;
+                        int gridX = grid[i].gridX + offset.x;
+                        int gridZ = grid[i].gridZ + offset.y;
 
                         if (gridX >= 0 && gridX < gridSizeX &&
                             gridZ >= 0 && gridZ < gridSizeZ)
                         {
-                            int neighborIndex = gridX + gridZ * gridSizeX;
-                            var yDistance = math.abs(currentCell.position.y - grid[neighborIndex].position.y);
+                            var yDistance = grid[i].position.y - grid[gridX + gridZ * grid.Length].position.y;
                             if (yDistance < maxHeightDifference)
-                            {
-                                allNeighbors[baseIndex + count] = neighborIndex;
                                 count++;
-                            }
                         }
                     }
                 }
+
+                neighborCounts[i] = count;
             }
         }
     }
