--- Assets/Scripts/NavigationGraph/Graph/Simple/SimpleGridNavigationGraph.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 70a5903f720d44c49f83f673a252a503
-timeCreated: 1738701914
