+++ Assets/Scripts/NavigationGraph/GraphFactory.cs
@@ -1,5 +1,5 @@
-using NavigationGraph.Graph;
-using NavigationGraph.RaycastCheck;
+using NavigationGraph.Graph.Layered;
+using NavigationGraph.Graph.Planar;
 
 namespace NavigationGraph
 {
@@ -9,8 +9,8 @@ namespace NavigationGraph
         {
             return graphType switch
             {
-                NavigationGraphType.Grid2D => new SimpleGridNavigationGraph(navigationGraphConfig),
-                NavigationGraphType.Grid3D => new WorldNavigationGraph(navigationGraphConfig),
+                NavigationGraphType.Grid2D => new PlanarGrid(navigationGraphConfig),
+                NavigationGraphType.Grid3D => new LayeredGrid(navigationGraphConfig),
                 _ => throw new System.NotImplementedException()
             };
         }
